Fix issuebox
